// src/types/sheet.type.ts
export interface TableData {
    ID: string;
    Monto: string;
    Apertura: string;
    Vencimiento: string;
    Tiempo: string;
    Intereses: string;
    Tasa: string;
    FechaPago?: string;
    Estado: string;
    Nombre: string;
    DNI: string;
    telefono: string;
    fechaNacimiento: string;
}